refactor(batch-exports): move the record batch queue out of spmc, delete it
Last of three PRs emptying `spmc.py`. What remained was the producer/consumer
handoff itself, which moves to `temporal/queue.py`: a queue bounded by bytes
rather than item count, plus the helpers for waiting on a producer and surfacing
its failures.

It sits flat in `temporal/` rather than under `pipeline/` because the callers are
overwhelmingly the destination workflows — 10 production call sites there against
3 in `pipeline/`. Putting it in the pipeline package would have meant every
destination importing from a package it isn't part of.

Also dropped `spmc.BatchExportField`, a byte-identical duplicate of
`service.BatchExportField`; `debug/debugger.py` was its only consumer and now
imports the original.

`spmc.py` is deleted. "SPMC" described the single-producer/multi-consumer design
that the internal-stage pipeline replaced, so the name outlived the thing it
named — the stale reference to it in `pipeline/consumer.py`'s docstring is fixed
too.

`test_spmc.py` split to follow its subjects: `test_queue.py`, `test_filters.py`,
`pipeline/test_query_ranges.py`, and the slicing tests into
`pipeline/test_producer.py`. All 12 tests kept.

Code moved verbatim; all six symbols verified byte-identical against master.

Generated-By: PostHog Code
Task-Id: e2c0e744-189b-4ac0-a1f4-c999b0a0e071
