+++ products/batch_exports/backend/temporal/destinations/databricks_batch_export.py
@@ -51,7 +51,7 @@ from products.batch_exports.backend.temporal.pipeline.transformer import (
     ParquetStreamTransformer,
 )
 from products.batch_exports.backend.temporal.pipeline.types import BatchExportResult
-from products.batch_exports.backend.temporal.spmc import RecordBatchQueue, wait_for_schema_or_producer
+from products.batch_exports.backend.temporal.queue import RecordBatchQueue, wait_for_schema_or_producer
 from products.batch_exports.backend.temporal.utils import (
     JsonType,
     handle_non_retryable_errors,
