+++ products/batch_exports/backend/temporal/pipeline/producer.py
@@ -14,7 +14,7 @@ from posthog.temporal.common.logger import get_write_only_logger
 
 from products.batch_exports.backend.temporal.metrics import CumulativeTimer
 from products.batch_exports.backend.temporal.pipeline.internal_stage import get_base_s3_staging_folder, get_s3_client
-from products.batch_exports.backend.temporal.spmc import RecordBatchQueue
+from products.batch_exports.backend.temporal.queue import RecordBatchQueue
 from products.batch_exports.backend.temporal.utils import make_retryable_with_exponential_backoff
 
 if typing.TYPE_CHECKING:
